+++ src/routers/auth.py
@@ -18,6 +18,7 @@ router_auth = APIRouter(prefix="/auth", tags=["Auth"])
 get_refresh_token = HTTPBearer()
 
 
+
 @router_auth.post(
     "/register", response_model=UserResponse, status_code=status.HTTP_201_CREATED
 )
@@ -121,15 +122,26 @@ async def request_email(body: RequestEmail, background_tasks: BackgroundTasks, r
     return {"message": "Check your email for confirmation."}
 
 
+from fastapi import HTTPException, status
+
+
+
+
 @router_auth.post("/password-reset", response_model=None)
 async def request_password_reset(password_reset_request: RequestEmail, db: AsyncSession = Depends(get_db)):
     try:
         print("Before reset_password_and_notify_user")
-        await auth_service.reset_password_and_notify_user(password_reset_request.email, db)
+
+        # Перевірка наявності користувача
+        user = await auth_service.get_user_by_email(password_reset_request.email, db)
+        if not user:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+
+        new_password = auth_service.generate_random_password()
+        await auth_service.reset_password_and_notify_user(password_reset_request.email, new_password, db)
         print("After reset_password_and_notify_user")
         return {"message": "Password reset request successful. Check your email for the new password."}
     except Exception as e:
         print(f"Error: {e}")
         raise HTTPException(status_code=500, detail=str(e))
-
